Remove surflock as the module
