--- buildconfig/stubs/pygame/surflock.pyi
@@ -1,2 +1,0 @@
-# surflock is a private pygame module that does not export any public API
-# this file is kept here to make stubtest happy
