+++ src_c/_pygame.h
@@ -615,8 +615,7 @@ typedef enum {
 #define PYGAMEAPI_RECT_NUMSLOTS 10
 #define PYGAMEAPI_JOYSTICK_NUMSLOTS 3
 #define PYGAMEAPI_DISPLAY_NUMSLOTS 2
-#define PYGAMEAPI_SURFACE_NUMSLOTS 4
-#define PYGAMEAPI_SURFLOCK_NUMSLOTS 6
+#define PYGAMEAPI_SURFACE_NUMSLOTS 10
 #define PYGAMEAPI_RWOBJECT_NUMSLOTS 5
 #define PYGAMEAPI_PIXELARRAY_NUMSLOTS 2
 #define PYGAMEAPI_COLOR_NUMSLOTS 5
