+++ src_c/meson.build
@@ -134,15 +134,6 @@ surface = py.extension_module(
 )
 endif
 
-surflock = py.extension_module(
-    'surflock',
-    'surflock.c',
-    c_args: warnings_error,
-    dependencies: pg_base_deps,
-    install: true,
-    subdir: pg,
-)
-
 time = py.extension_module(
     'time',
     'time.c',
