+++ src_py/__init__.py
@@ -146,7 +146,6 @@ from pygame.constants import *  # now has __all__ pylint: disable=wildcard-impor
 from pygame.version import *  # pylint: disable=wildcard-import; lgtm[py/polluting-import]
 from pygame.rect import Rect, FRect
 from pygame.rwobject import encode_string, encode_file_path
-import pygame.surflock
 import pygame.color
 
 Color = pygame.color.Color
